refactor(server): split prometheus metrics renderers
